Changed concrete classes to interfaces to make them more amenable to spring autowiring
