+++ src/main/java/org/owasp/dependencytrack/service/LibraryVersionService.java
@@ -1,132 +1,65 @@
-/*
- * This file is part of Dependency-Track.
- *
- * Dependency-Track is free software: you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * Dependency-Track is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * Dependency-Track. If not, see http://www.gnu.org/licenses/.
- *
- * Copyright (c) Axway. All Rights Reserved.
- */
-
 package org.owasp.dependencytrack.service;
 
-import org.owasp.dependencytrack.dao.LibraryVersionDao;
-import org.owasp.dependencytrack.model.ApplicationVersion;
-import org.owasp.dependencytrack.model.Library;
-import org.owasp.dependencytrack.model.LibraryVendor;
-import org.owasp.dependencytrack.model.LibraryVersion;
-import org.owasp.dependencytrack.model.License;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.owasp.dependencytrack.model.*;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
-@Service
-public class LibraryVersionService {
-
-    @Autowired
-    private LibraryVersionDao libraryVersionDao;
-
+/**
+ * Created by Jason Wraxall on 1/12/15.
+ */
+public interface LibraryVersionService {
+    @Transactional
+    List<LibraryVersion> getDependencies(ApplicationVersion version);
 
     @Transactional
-    public List<LibraryVersion> getDependencies(ApplicationVersion version) {
-        return libraryVersionDao.getDependencies(version);
-    }
+    void addDependency(int appversionid, int libversionid);
 
     @Transactional
-    public void addDependency(int appversionid, int libversionid) {
-        libraryVersionDao.addDependency(appversionid, libversionid);
-    }
-
-    @Transactional
-    public void deleteDependency(int appversionid, int libversionid) {
-        libraryVersionDao.deleteDependency(appversionid, libversionid);
-    }
+    void deleteDependency(int appversionid, int libversionid);
 
     /*
-        Returns a List of all LibraryVendors available in the application along with all child objects
-     */
+            Returns a List of all LibraryVendors available in the application along with all child objects
+         */
     @Transactional
-    public List<LibraryVendor> getLibraryHierarchy() {
-        return libraryVersionDao.getLibraryHierarchy();
-    }
-
-
+    List<LibraryVendor> getLibraryHierarchy();
 
     @Transactional
-    public void updateLibrary(int vendorid, int licenseid, int libraryid,
-                              int libraryversionid, String libraryname, String libraryversion,
-                              String vendor, String license,  String language) {
-
-        libraryVersionDao.updateLibrary(vendorid, licenseid, libraryid,
-                libraryversionid, libraryname, libraryversion, vendor, license, language);
-    }
+    void updateLibrary(int vendorid, int licenseid, int libraryid,
+                       int libraryversionid, String libraryname, String libraryversion,
+                       String vendor, String license, String language);
 
     @Transactional
-    public void removeLibrary(Integer id) {
-        libraryVersionDao.removeLibrary(id);
-    }
+    void removeLibrary(Integer id);
 
     @Transactional
-    public List<License> listLicense(Integer id) {
-        return libraryVersionDao.listLicense(id);
-    }
+    List<License> listLicense(Integer id);
 
     @Transactional
-    public List<LibraryVersion> allLibrary() {
-        return libraryVersionDao.allLibrary();
-    }
+    List<LibraryVersion> allLibrary();
 
     @Transactional
-    public List<Library> uniqueLibrary() {
-        return libraryVersionDao.uniqueLibrary();
-    }
+    List<Library> uniqueLibrary();
 
     @Transactional
-    public List<License> uniqueLicense() {
-        return libraryVersionDao.uniqueLicense();
-    }
+    List<License> uniqueLicense();
 
     @Transactional
-    public List<LibraryVendor> uniqueVendor() {
-        return libraryVersionDao.uniqueVendor();
-    }
+    List<LibraryVendor> uniqueVendor();
 
     @Transactional
-    public List<String> uniqueLang() {
-        return libraryVersionDao.uniqueLang();
-    }
+    List<String> uniqueLang();
 
     @Transactional
-    public List<String> uniqueVer() {
-        return libraryVersionDao.uniqueVer();
-    }
+    List<String> uniqueVer();
 
     @Transactional
-    public void addLibraries(String libraryname, String libraryversion, String vendor, String license, MultipartFile file, String language) {
-
-        libraryVersionDao.addLibraries(libraryname, libraryversion, vendor,  license,  file,  language);
-    }
+    void addLibraries(String libraryname, String libraryversion, String vendor, String license, MultipartFile file, String language);
 
     @Transactional
-    public void uploadLicense(int licenseid, MultipartFile file, String editlicensename) {
-        libraryVersionDao.uploadLicense(licenseid, file, editlicensename);
-    }
+    void uploadLicense(int licenseid, MultipartFile file, String editlicensename);
 
     @Transactional
-    public List<LibraryVersion> keywordSearchLibraries(String searchTerm) {
-        return libraryVersionDao.keywordSearchLibraries(searchTerm);
-    }
-
+    List<LibraryVersion> keywordSearchLibraries(String searchTerm);
 }
