+++ src/main/java/org/owasp/dependencytrack/service/ReportService.java
@@ -1,39 +1,12 @@
-/*
- * This file is part of Dependency-Track.
- *
- * Dependency-Track is free software: you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * Dependency-Track is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * Dependency-Track. If not, see http://www.gnu.org/licenses/.
- *
- * Copyright (c) Axway. All Rights Reserved.
- */
-
 package org.owasp.dependencytrack.service;
 
 import org.owasp.dependencycheck.reporting.ReportGenerator;
-import org.owasp.dependencytrack.dao.ReportDao;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Service
-public class ReportService {
-
-    @Autowired
-    private ReportDao reportDao;
-
+/**
+ * Created by Jason Wraxall on 1/12/15.
+ */
+public interface ReportService {
     @Transactional
-    public String generateDependencyCheckReport(int applicationVersionId, ReportGenerator.Format format) {
-        return reportDao.generateDependencyCheckReport(applicationVersionId, format);
-    }
-
+    String generateDependencyCheckReport(int applicationVersionId, ReportGenerator.Format format);
 }
